+++ tests/testthat/test-binary_endpoints.R
@@ -98,3 +98,63 @@ test_that("`s_odds_ratio` works also when there are only 2 `col_by` levels", {
   )
   expect_equal(result, expected, tol = 0.001)
 })
+
+test_that("`s_proportion` works with Agresti-Coull CI", {
+  # "Mid" case.
+  rsp1 <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
+  result1 <- s_proportion(
+    x = rsp1,
+    conf_level = 0.9,
+    prop_ci_method = "agresti-coull"
+  )
+  expected1 <- list(
+    prop = 4 / 6,
+    prop_ci = c(0.3436, 0.8862),  # From SAS.
+    label_ci = "90% CI for Response Rates (Agresti-Coull)"
+  )
+  expect_equal(result1, expected1, tol = 0.0001)
+
+  # Corner case: Only responders.
+  rsp2 <- c(TRUE, TRUE, TRUE, TRUE)
+  result2 <- s_proportion(
+    x = rsp2,
+    conf_level = 0.95,
+    prop_ci_method = "agresti-coull"
+  )
+  expected2 <- list(
+    prop = 1,
+    prop_ci = c(0.4540, 1),  # From SAS.
+    label_ci = "95% CI for Response Rates (Agresti-Coull)"
+  )
+  expect_equal(result2, expected2, tol = 0.00011)  # Small additional difference here.
+})
+
+test_that("`s_proportion` works with Jeffreys CI", {
+  # "Mid" case.
+  rsp1 <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
+  result1 <- s_proportion(
+    x = rsp1,
+    conf_level = 0.9,
+    prop_ci_method = "jeffreys"
+  )
+  expected1 <- list(
+    prop = 4 / 6,
+    prop_ci = c(0.3408, 0.8957),  # From SAS.
+    label_ci = "90% CI for Response Rates (Jeffreys)"
+  )
+  expect_equal(result1, expected1, tol = 0.0001)
+
+  # Corner case: Only responders.
+  rsp2 <- c(TRUE, TRUE, TRUE, TRUE)
+  result2 <- s_proportion(
+    x = rsp2,
+    conf_level = 0.95,
+    prop_ci_method = "jeffreys"
+  )
+  expected2 <- list(
+    prop = 1,
+    prop_ci = c(0.5552, 1),  # From SAS.
+    label_ci = "95% CI for Response Rates (Jeffreys)"
+  )
+  expect_equal(result2, expected2, tol = 0.0001)
+})